pyproject.toml
[tool.poetry]
name = "strawberry-django-social-auth"
version = "0.1.0"
description = "Graphql Social authentication system with Strawberry for Django."
license = "MIT"
authors = ["Shritesh Jamulkar<shritesh.sj@gmail.com>"]
maintainers = ["Shritesh Jamulkar <shritesh.sj@gmail.com>"]
readme = "README.md"
classifiers = [
    'Environment :: Web Environment',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Framework :: Django',
    'Framework :: Django :: 3.2',
    'Framework :: Django :: 4.1',
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11"
]
packages = [
    { include = "gql_social_auth" }
]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
Django = ">=3.2,<4.3"
PyJWT = ">=2.6.0,<3.0"
strawberry-graphql-django = ">=0.10.5"
social-auth-app-django= ">=2.1.0"
strawberry-django-auth = ">=0.373.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
coverage = "^7.2"
pytest = "^7.2"
pytest-cov = "^4.0"
types-cryptography = "^3.3.23"
django-mock-queries = "^2.1.7"
types-mock = "^5.0.0"
types-jwt = "^0.1.0"
types-pkg-resources = "^0.1.0"
mkdocs = ">=1.3.0"
mkdocs-markdownextradata-plugin = ">=0.2.5"
mkdocs-material = ">=8.5.4"
mkdocs-minify-plugin = ">=0.5.0"
pymdown-extensions = ">=9.9"
Markdown = ">=3.4.1"
Faker = ">=18.2.0"
pytest-asyncio = ">=0.21.0"
pydoc-markdown = {extras = ["novella"], version = ">=4.6.4"}
setuptools = ">=67.6.0"
channels = {extras = ["daphne"], version = ">=4.0.0"}
pytest-django = "^4.5.2"

[tool.pytest.ini_options]
django_find_project = false
addopts = "-ra -s -p no:warnings"
testpaths = [
    "tests",
]
DJANGO_SETTINGS_MODULE = "testproject.settings"
python_files = "tests.py test_*.py *_tests.py"

[tool.poetry.urls]
"Homepage" = "https://github.com/shritesh99/strawberry-django-social-auth"
"Documentation" = "https://github.com/shritesh99/strawberry-django-social-auth"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | __pycached__
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
force_sort_within_sections = true

[tool.mypy]
plugins = ["mypy_django_plugin.main", "strawberry.ext.mypy_plugin"]
ignore_missing_imports = "True"

[tool.django-stubs]
django_settings_module = "testproject.settings"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.ruff]
line-length = 100
select = ["E", "F", "ERA", "I"]
ignore = [
    "TID252",
    "E501",  # line too long, handled by black
]
fix = true

src = ["gql_social_auth", "tests", "testproject"]